% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{example_liv}
\alias{example_liv}
\title{Example L-I-V sweep data}
\format{
\subsection{\code{example_liv}}{

A data frame with 500 rows and 4 columns:
\describe{
\item{current[mA]}{source current (mA)}
\item{power[mW]}{measured output power (mW)}
\item{voltage[mA]}{measured voltage (V)}
\item{mpd_current[mA]}{measured monitor photodiode current (mA)}
...
}
}
}
\source{
internal measurement csv file
}
\usage{
example_liv
}
\description{
This is an example LIV measurement of a high power C-band DFB.
}
\keyword{datasets}
