% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bw20dB.R
\name{extract_bw20dB}
\alias{extract_bw20dB}
\title{Extract the signal bandwidth, at 20dB down, from spectrum data}
\usage{
extract_bw20dB(wav, power)
}
\arguments{
\item{wav}{A numeric vector of spectrum wavelength data}

\item{power}{A numeric vector of power spectrum data}
}
\value{
A numeric value, or NA if the calculation failed
}
\description{
Extract the signal bandwidth, at 20dB down, from spectrum data
}
\examples{
wav = seq(1300, 1320, by = 0.1)
pow = rep(-70, length(wav)) + rnorm(length(wav), 0, 1)
pow[101:103] = c(-30, -5, -30)
pow[50:52] = c(-40, -38, -40)
bandwidth = extract_bw20dB(wav, pow)
print(bandwidth)
}
