% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eml_dcer.R
\name{extract_eml_dcer}
\alias{extract_eml_dcer}
\title{Extract EML DC extinction ratio from P-Vea data}
\usage{
extract_eml_dcer(voltage, power, V0 = 0, V1 = -1)
}
\arguments{
\item{voltage}{A numeric vector of source voltage values}

\item{power}{A numeric vector of measure optical power values}

\item{V0}{A numeric value for the reference voltage. The devault value is V0
= 0V}

\item{V1}{A numeric value for the applied EAM voltage which attenuates the
laser power. The default value is V1 = -1.0V}
}
\value{
A numeric value DCER = 10*log10(P0/P1)
}
\description{
Extract EML DC extinction ratio from P-Vea data
}
