% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eml_dpdv_max.R
\name{extract_eml_dpdv_max}
\alias{extract_eml_dpdv_max}
\title{Extract the max dP/dV from the EAM P-Vea curve}
\usage{
extract_eml_dpdv_max(V, P)
}
\arguments{
\item{V}{a numeric vector of sourced EAM voltages}

\item{P}{a numeric vector of measured EML output powers}
}
\value{
a numeric value
}
\description{
Extract the max dP/dV from the EAM P-Vea curve
}
