% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eml_pf.R
\name{extract_eml_pf}
\alias{extract_eml_pf}
\title{Extract Pf from EML P-Vea data}
\usage{
extract_eml_pf(voltage, power, V0 = 0)
}
\arguments{
\item{voltage}{A numeric vector of source voltages}

\item{power}{A numeric vector of measured optical powers}

\item{V0}{A numeric value of modulator bias voltage}
}
\value{
A numeric value for Pf
}
\description{
Extract Pf from EML P-Vea data
}
