% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eml_rphoto.R
\name{extract_eml_rphoto}
\alias{extract_eml_rphoto}
\title{Extract EAM photocurrent resistance}
\usage{
extract_eml_rphoto(voltage, dVdI, V0)
}
\arguments{
\item{voltage}{A numeric vector of sourced EA voltages}

\item{dVdI}{A numeric vector of calculated dVdI values}

\item{V0}{A numeric value of the extraction voltage}
}
\value{
A numeric value of the photocurrent resistance at the operating voltage
}
\description{
Extract EAM photocurrent resistance
}
