% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eml_vop.R
\name{extract_eml_vop}
\alias{extract_eml_vop}
\title{Extract EAM operating voltage}
\usage{
extract_eml_vop(voltage, dPdV)
}
\arguments{
\item{voltage}{A numeric vector of sourced EA voltages}

\item{dPdV}{A numeric vector of dPdV values}
}
\value{
A numeric value of the operating voltage
}
\description{
Extract EAM operating voltage
}
