% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eml_vop2.R
\name{extract_eml_vop2}
\alias{extract_eml_vop2}
\title{New one}
\usage{
extract_eml_vop2(voltage, dPdV)
}
\arguments{
\item{voltage}{a numeric vector}

\item{dPdV}{a numeric vector}
}
\value{
a numeric value
}
\description{
New one
}
