% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_full_width.R
\name{extract_full_width}
\alias{extract_full_width}
\title{Extract full-width from a distribution}
\usage{
extract_full_width(x, y, threshold = -20, plot_debug = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{threshold}{a numeric value}

\item{plot_debug}{a Boolean value}
}
\description{
Extract full-width from a distribution
}
