% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_iop_from_pi.R
\name{extract_iop_from_pi}
\alias{extract_iop_from_pi}
\title{Extract LD operating current from P-I data}
\usage{
extract_iop_from_pi(I, P, Pop = 0.01)
}
\arguments{
\item{I}{Sweep currents}

\item{P}{Sweep powers}

\item{Pop}{The operating power to extract Iop at}
}
\value{
A numeric value
}
\description{
Extract LD operating current from P-I data
}
\examples{
I = seq(0, 10)
P = sqrt(seq(0, 10))
Pop = 1
Iop = extract_iop_from_pi(I, P, Pop = Pop)
plot(I, P, type = "l")
points(Iop, Pop, col = "red")
}
