% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_isat_from_pi.R
\name{extract_isat_from_pi}
\alias{extract_isat_from_pi}
\title{Extract saturation current (Isat) from power-current data}
\usage{
extract_isat_from_pi(I, P)
}
\arguments{
\item{I}{A numeric vector of sweep currents}

\item{P}{A numeric vector of power measurements}
}
\value{
A numeric value
}
\description{
Extract saturation current (Isat) from power-current data
}
