% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ith_from_pi.R
\name{extract_ith_from_pi}
\alias{extract_ith_from_pi}
\title{Extract threshold current from P-I data}
\usage{
extract_ith_from_pi(
  I,
  P,
  method = "intercept",
  n1_smooth = 1,
  n2_smooth = 1,
  n3_smooth = 1
)
}
\arguments{
\item{I}{A numeric vector of currents}

\item{P}{A numeric vector of powers}

\item{method}{The threshold current extraction method}

\item{n1_smooth}{An integer. P-I curve moving average filter length.}

\item{n2_smooth}{An integer. dP/dI curve moving average filter length.}

\item{n3_smooth}{An integer. d2P/dI2 curve moving average filter length.}
}
\value{
A numeric value
}
\description{
Extract threshold current from P-I data
}
\examples{
I = example_liv$`current[mA]`
P = example_liv$`power[mW]`
Ith = extract_ith_from_pi(I, P, method = "second_derivative")
}
