% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_kink_from_pi.R
\name{extract_kink_from_pi}
\alias{extract_kink_from_pi}
\title{Extract kink deviation from P-I curve}
\usage{
extract_kink_from_pi(I, P, Istart, Istop, plot_debug = FALSE)
}
\arguments{
\item{I}{a numeric vector of sourced current values}

\item{P}{a numeric vector of measured output powers}

\item{Istart}{a numeric value for the beginning of the current span to check}

\item{Istop}{a numeric value for the end of the current span to check}

\item{plot_debug}{a logical value to enable plot debug mode}
}
\value{
a data frame
}
\description{
Extract kink deviation from P-I curve
}
