% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mode_hop.R
\name{extract_mode_hop}
\alias{extract_mode_hop}
\title{Mode hop detection analysis of peak wavelength vs bias current data}
\usage{
extract_mode_hop(If, Lp, plot_debug = FALSE, plot_title = "default_title")
}
\arguments{
\item{If}{Numeric vector of bias currents}

\item{Lp}{Numeric vector of peak wavelengths}

\item{plot_debug}{Logical flag to enable visual inspection of the analysis}

\item{plot_title}{String for the debug plot title}
}
\value{
A data frame with analysis metrics
}
\description{
Mode hop detection analysis of peak wavelength vs bias current data
}
