% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pce.R
\name{extract_pce}
\alias{extract_pce}
\title{Extract Power Conversion Efficiency}
\usage{
extract_pce(I, P, V, I0)
}
\arguments{
\item{I}{sweep currents}

\item{P}{sweep powers}

\item{V}{sweep voltages}

\item{I0}{current to extract the parameter at}
}
\value{
a numeric value as percentage
}
\description{
Extract Power Conversion Efficiency
}
