% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_peak_power.R
\name{extract_peak_power}
\alias{extract_peak_power}
\title{Extract peak power from optical spectrum}
\usage{
extract_peak_power(power)
}
\arguments{
\item{power}{A numeric vector of spectrum power}
}
\value{
A numeric value
}
\description{
Extract peak power from optical spectrum
}
