% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pf_from_pi.R
\name{extract_pf_from_pi}
\alias{extract_pf_from_pi}
\title{Extract the output power at a given bias current}
\usage{
extract_pf_from_pi(I, P, I0)
}
\arguments{
\item{I}{Sweep currents}

\item{P}{Sweep powers}

\item{I0}{A bias current to extract output power at}
}
\value{
A numeric value
}
\description{
Extract the output power at a given bias current
}
\examples{
I = seq(0, 10)
P = sqrt(seq(0, 5, length.out = length(I)))
I0 = 3.75
Pf = extract_pf_from_pi(I, P, I0 = I0)
}
