% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rs_from_vi.R
\name{extract_rs_from_vi}
\alias{extract_rs_from_vi}
\title{Extract series resistance from V-I data}
\usage{
extract_rs_from_vi(I, V, I0, Ispan = 4 * 0.001)
}
\arguments{
\item{I}{Vector of currents}

\item{V}{Vector of voltages}

\item{I0}{A bias current to extract the series resistance at}

\item{Ispan}{The span width to apply a linear fit to}
}
\value{
A numeric value
}
\description{
Extract series resistance from V-I data
}
\examples{
I = seq(0, 100e-3, 1e-3)
V = 1 + seq(0, 1, length.out = length(I))
I0 = 50e-3
Ispan = 6e-3
Rs = extract_rs_from_vi(I, V, I0 = I0, Ispan = Ispan)
print(Rs)
}
