% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rth.R
\name{extract_rth}
\alias{extract_rth}
\title{Extract thermal resistance using multi-temperature method}
\usage{
extract_rth(temps, pow_thermal, peak_wav)
}
\arguments{
\item{temps}{A vector of temperature values}

\item{pow_thermal}{A vector of thermal power dissipation values}

\item{peak_wav}{A vector of peak wavelength values}
}
\value{
A data frame with column names Rth, mRthhs, mRthPth, b, and Pthr2
}
\description{
Extract thermal resistance using multi-temperature method
}
