% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rth2.R
\name{extract_rth2}
\alias{extract_rth2}
\title{Extract thermal resistance using single-temperature method}
\usage{
extract_rth2(Pth, Lp, m_hs)
}
\arguments{
\item{Pth}{A vector of thermal power dissipation values}

\item{Lp}{A vector of peak wavelength values}

\item{m_hs}{A numeric value of the wavelength temperature coefficient (nm/C)}
}
\value{
A data frame with column names Rth2, mRth2Pth, Rth2r2
}
\description{
Extract thermal resistance using single-temperature method
}
