% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_se_from_pi.R
\name{extract_se_from_pi}
\alias{extract_se_from_pi}
\title{Extract slope efficiency at a given bias current}
\usage{
extract_se_from_pi(I, P, I0, Ispan = 0.004)
}
\arguments{
\item{I}{Sweep currents}

\item{P}{Sweep output powers}

\item{I0}{Bias current to extract slope efficiency at}

\item{Ispan}{The current span width that a linear fit is applied to}
}
\value{
A numeric vector
}
\description{
Extract slope efficiency at a given bias current
}
\examples{
I = seq(0, 10)
P = sqrt(seq(0, 5, length.out = length(I)))
I0 = 5
SE = extract_se_from_pi(I, P, I0 = I0, Ispan = 2)
}
