% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_smsr.R
\name{extract_smsr}
\alias{extract_smsr}
\title{Extract side mode suppresion ratio from spectrum data}
\usage{
extract_smsr(wav, power, smooth_fraction = 0)
}
\arguments{
\item{wav}{A vector of spectrum wavelength data}

\item{power}{A vector of spectrum power data}

\item{smooth_fraction}{A parameter for smoothing the power data.
smooth_fraction = smooth_points/total_points. smooth_fraction > 0.003 is not
recommended.}
}
\value{
A numeric value
}
\description{
Extract side mode suppresion ratio from spectrum data
}
\examples{
wav = seq(1300, 1320, by = 1)
pow = rep(-70, length(wav))
pow[6] = -45
pow[10] = -10
smsr = extract_smsr(wav, pow)
print(smsr)
}
