% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_smsr2.R
\name{extract_smsr2}
\alias{extract_smsr2}
\title{Extract side mode suppression ratio from spectrum data using masking method}
\usage{
extract_smsr2(wav, power, mask_width = 2)
}
\arguments{
\item{wav}{A numeric vector of spectrum wavelength data}

\item{power}{A numeric vector of spectrum power data}

\item{mask_width}{A numeric value, in nanometers (nm), defining the mask width}
}
\value{
A numeric value
}
\description{
Extract side mode suppression ratio from spectrum data using masking method
}
\examples{
wav = seq(1300, 1320, by = 0.1)
pow = rep(-70, length(wav))
pow[wav == 1310] = -10
pow[wav == 1305] = -45
smsr = extract_smsr2(wav, pow, mask_width = 2.0)
print(smsr)
}
