% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vf_from_vi.R
\name{extract_vf_from_vi}
\alias{extract_vf_from_vi}
\title{Extract forward voltage from V-I data}
\usage{
extract_vf_from_vi(I, V, I0)
}
\arguments{
\item{I}{Sweep currents}

\item{V}{Sweep voltages}

\item{I0}{Bias current to extract forward voltage at}
}
\value{
A numeric value
}
\description{
Extract forward voltage from V-I data
}
\examples{
I = seq(0, 10)
V = sqrt(seq(0, 10))
I0 = 4
Vf = extract_vf_from_vi(I, V, I0 = I0)
plot(I, V, type = "l")
points(I0, Vf, col = "red")
}
