% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dldt.R
\name{extract_wl_temp_coef}
\alias{extract_wl_temp_coef}
\title{Extract wavelength temperature coefficient}
\usage{
extract_wl_temp_coef(temperature, peak_wl)
}
\arguments{
\item{temperature}{A vector of temperature data}

\item{peak_wl}{A vector of peak wavelength data}
}
\value{
A numeric value
}
\description{
Extract wavelength temperature coefficient
}
\examples{
T = c(25, 50, 75)
Lp = c(1308, 1310.4, 1312.3)
temp_coef = extract_wl_temp_coef(T, Lp)$temp_coef
r2 = extract_wl_temp_coef(T, Lp)$rsquared
print(temp_coef)
}
