% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractI_from_IV.R
\name{extractI_from_IV}
\alias{extractI_from_IV}
\title{Extract current at a given voltage}
\usage{
extractI_from_IV(V, I, V0 = -1)
}
\arguments{
\item{V}{Sweep voltages}

\item{I}{Sweep currents}

\item{V0}{Voltage to extract current at. It is usually a single value, but it
can also be a vector}
}
\value{
A numeric vector
}
\description{
Extract current at a given voltage
}
\examples{
V = c(0, 1, 2, 3)
I = c(-5, -2, 2, 5)
V0 = 1.5
I = extractI_from_IV(V, I, V0 = V0)
}
