% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_max.R
\name{interp_max}
\alias{interp_max}
\title{Interpolate maximum value of x-y data}
\usage{
interp_max(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numeric value
}
\description{
Interpolate maximum value of x-y data
}
