% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnt_outlier_tukey.R
\name{isnt_outlier_tukey}
\alias{isnt_outlier_tukey}
\title{Determine outliers using Tukey's rule}
\usage{
isnt_outlier_tukey(x, k = 1.5, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{k}{A numeric value, default is k = 1.5}

\item{na.rm}{A logical indicating whether NA values should be removed before
the computation proceeds}
}
\value{
A logical vector, TRUE means that the value is not an outlier
}
\description{
Determine outliers using Tukey's rule
}
