% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_osa_data_files.R
\name{load_raw_osa_data_files}
\alias{load_raw_osa_data_files}
\title{Load many raw LIV data files into a data frame}
\usage{
load_raw_osa_data_files(paths, pattern = "-OSA.csv$")
}
\arguments{
\item{paths}{The file path(s) to folder(s) containing the csv files}

\item{pattern}{A pattern to match certain file names}
}
\value{
A data frame with all the rows of data combined
}
\description{
Load many raw LIV data files into a data frame
}
