% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_linear_fit.R
\name{my_linear_fit}
\alias{my_linear_fit}
\title{Create linear fit to x-y data}
\usage{
my_linear_fit(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
A data frame
}
\description{
Create linear fit to x-y data
}
\examples{
x = 1:3
y = 2*x + 5 + rnorm(length(x), mean = 0 , sd = 0.1)
my_linear_fit(x, y)
}
