% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_smooth.R
\name{my_smooth}
\alias{my_smooth}
\title{Smooth data using moving averaging}
\usage{
my_smooth(x, N)
}
\arguments{
\item{x}{A numeric vector to be smoothed}

\item{N}{An integer that defines the span of the moving average, it should be
an odd number.}
}
\value{
A numeric vector
}
\description{
Smooth data using moving averaging
}
\examples{
x = rnorm(100, 0, 1)
y = my_smooth(x, N = 9)
plot(x)
lines(y, col = "red")
}
