% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary_stats.R
\name{my_summary_stats}
\alias{my_summary_stats}
\title{My summary statistics function}
\usage{
my_summary_stats(df)
}
\arguments{
\item{df}{A data frame with data in column x}
}
\value{
A data frame with summary statistics
}
\description{
My summary statistics function
}
