% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_raw_eml_data.R
\name{summarize_raw_eml_data}
\alias{summarize_raw_eml_data}
\title{Create a summary from the raw data}
\usage{
summarize_raw_eml_data(df_eml)
}
\arguments{
\item{df_eml}{a data frame}
}
\value{
a data frame
}
\description{
Create a summary from the raw data
}
