% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_raw_osa_data.R
\name{summarize_raw_osa_data}
\alias{summarize_raw_osa_data}
\title{Create a summary from optical spectra raw data}
\usage{
summarize_raw_osa_data(df_osa)
}
\arguments{
\item{df_osa}{a data frame containing wavelength and power data}
}
\value{
a data frame containing OSA-extracted parameteric data
}
\description{
Create a summary from optical spectra raw data
}
