% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ith_from_pi.R
\name{extract_ith_from_pi}
\alias{extract_ith_from_pi}
\title{Extract threshold current}
\usage{
extract_ith_from_pi(
  I,
  P,
  n1_smooth = 1,
  n2_smooth = 1,
  n3_smooth = 1,
  plot_debug = FALSE
)
}
\arguments{
\item{I}{A numeric vector of sourced current values}

\item{P}{A numeric vector of measured power values}

\item{n1_smooth}{Smooth span for P-I curve}

\item{n2_smooth}{Smooth span for P-I curve 1st derivative}

\item{n3_smooth}{Smooth span for P-I curve 2nd derivative}

\item{plot_debug}{A Boolean value for setting the plot inspection mode}
}
\description{
Extract threshold current
}
